pyproject.toml
[project]
name = "thsr-ticket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4==4.14.3",
    "ddddocr==1.5.6",
    "jsonschema==4.26.0",
    "numpy==2.2.6",
    "pillow==12.1.1",
    "pydantic==2.12.5",
    "requests==2.32.5",
]
